Add files via upload
after many hours of tinkering with html div manipulation, I finally got the output main box to stop expanding when text is printed into it. Working on parsing the input and switching from different input modes like when the game presents an npc dialogue, or a trade, or other types of indexed options.